+++ fractals/src/main/java/me/catmousedog/fractals/fractals/Fractal.java
@@ -1,9 +1,16 @@
 package me.catmousedog.fractals.fractals;
 
+import javax.swing.JPanel;
+
+import me.catmousedog.fractals.ui.JPInterface;
+import me.catmousedog.fractals.ui.Savable;
+
 /**
- * represents a fractal, including its function for each point in 2D space
+ * Represents a fractal including its fractal function and colour filter. This
+ * class also holds a {@link Fractal#addFilter(JPanel)} which creates a
+ * {@link Savable} user interface.
  */
-public abstract class Fractal {
+public abstract class Fractal implements Savable {
 
 	protected int iterations = 100;
 
@@ -21,7 +28,30 @@ public abstract class Fractal {
 	 * @param y the y coordinate of the point
 	 * @return the value the fractal function returns
 	 */
-	public abstract double get(double x, double y);
+	public abstract Number get(double x, double y);
+
+	/**
+	 * Applies the calculated value 'v' to a colour filter and returns an integer
+	 * rgb value.
+	 * 
+	 * @param v value calculated by the fractal
+	 * @return the rgb value as an integer
+	 */
+	public abstract int filter(Number v);
+
+	/**
+	 * Adds all the necessary components to a {@link JPanel} on the
+	 * {@link JPInterface}. This includes any
+	 * 
+	 * @param jp
+	 */
+	public abstract void addFilter(JPanel jp);
+
+	/**
+	 * Returns the name of this fractal.
+	 */
+	@Override
+	public abstract String toString();
 
 	public void setIterations(int i) {
 		iterations = i;
@@ -30,4 +60,5 @@ public abstract class Fractal {
 	public int getIterations() {
 		return iterations;
 	}
+
 }