+++ fractals/src/main/java/me/catmousedog/fractals/fractals/Pixel.java
@@ -18,6 +18,6 @@ public class Pixel {
 	public double tx, ty;
 
 	// output
-	public double v;
+	public Number v;
 
 }