+++ fractals/src/main/java/me/catmousedog/fractals/fractals/concrete/TestFractal.java
@@ -1,0 +1,55 @@
+package me.catmousedog.fractals.fractals.concrete;
+
+import javax.swing.JPanel;
+
+import me.catmousedog.fractals.fractals.Fractal;
+import me.catmousedog.fractals.ui.Savable;
+
+public class TestFractal extends Fractal implements Savable {
+
+	@Override
+	public void save() {
+		// XXX Auto-generated method stub
+
+	}
+
+	@Override
+	public void update() {
+		// XXX Auto-generated method stub
+
+	}
+
+	@Override
+	public void preRender() {
+		// XXX Auto-generated method stub
+
+	}
+
+	@Override
+	public void postRender() {
+		// XXX Auto-generated method stub
+
+	}
+
+	@Override
+	public Number get(double x, double y) {
+		return 0;
+	}
+
+	@Override
+	public int filter(Number v) {
+		return 100000000;
+	}
+
+	@Override
+	public void addFilter(JPanel jp) {
+		// XXX Auto-generated method stub
+
+	}
+
+	@Override
+	public String toString() {
+		return "TestFractal";
+	}
+
+}